fix & refactor modals windows
