+++ src/components/decs/deck-dialogs/add-new-card-dialog/addNewCardDialog.module.scss
@@ -14,7 +14,7 @@
   width: 100%;
   height: 119px;
   margin: 20px 0 !important;
-  background: url('@/assets/jpg/cover-default.jpg') center / cover;
+  background: url('@/assets/webp/cover-default.webp') center / cover;
 }
 
 .svg {
