+++ src/components/decs/deck-dialogs/add-new-card-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './addNewCardDialog'
