+++ src/components/decs/deck-dialogs/add-new-deck-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './addNewDeckDialog'
