+++ src/components/decs/deck-dialogs/delete-card-dialog/deleteCardDialog.module.scss
