+++ src/components/decs/deck-dialogs/delete-card-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './deleteCardDialog'
