--- src/components/ui/modal/addNewCard/index.ts
@@ -1,1 +1,0 @@
-export * from './modalAddNewCard'
