--- src/components/ui/modal/addNewDesk/index.ts
@@ -1,1 +1,0 @@
-export * from './modalAddNewDesk'
