--- src/components/ui/modal/deleteCard/index.ts
@@ -1,1 +1,0 @@
-export * from './deleteCard'
